requirements.txt
appnope==0.1.3
arrow==1.2.3
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
cssselect==1.2.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
fake-useragent==1.1.3
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.11.0
idna==3.4
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jieba3k==0.35.1
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
lxml==4.9.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.1.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4
stack-data==0.6.2
tinysegmenter==0.3
tldextract==3.4.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
urllib3==1.26.14
user-agent==0.1.10
wcwidth==0.2.6
